+++ full-moon/src/tokenizer/structs.rs
@@ -4,14 +4,13 @@ use crate::{
     ShortString,
 };
 
+use super::{Lexer, LexerResult};
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
 use std::{
     cmp::Ordering,
     fmt::{self, Display},
 };
-
-use super::{Lexer, LexerResult};
 
 macro_rules! symbol {
     {
@@ -123,19 +122,26 @@ symbol! {
 
         [lua52 | luajit] Goto => "goto",
 
-        [luau] PlusEqual => "+=",
-        [luau] MinusEqual => "-=",
-        [luau] StarEqual => "*=",
-        [luau] SlashEqual => "/=",
+        [luau | cfxlua] PlusEqual => "+=",
+        [luau | cfxlua] MinusEqual => "-=",
+        [luau | cfxlua] StarEqual => "*=",
+        [luau | cfxlua] SlashEqual => "/=",
+
         [luau] DoubleSlashEqual => "//=",
         [luau] PercentEqual => "%=",
-        [luau] CaretEqual => "^=",
+        [luau | cfxlua] CaretEqual => "^=",
         [luau] TwoDotsEqual => "..=",
 
         [luau | lua53] Ampersand => "&",
         [luau] ThinArrow => "->",
         [luau | lua52 | luajit] TwoColons => "::",
         [luau] AtSign => "@",
+
+        [cfxlua] DoubleLessThanEqual => "<<=",
+        [cfxlua] DoubleGreaterThanEqual => ">>=",
+        [cfxlua] AmpersandEqual => "&=",
+        [cfxlua] PipeEqual => "|=",
+        [cfxlua] QuestionMarkDot => "?.",
 
         Caret => "^",
         Colon => ":",
@@ -306,19 +312,37 @@ pub enum TokenType {
         /// If it is the beginning, middle, end, or a standalone string.
         kind: InterpolatedStringKind,
     },
+
+    #[cfg(feature = "cfxlua")]
+    /// A single/multi line C-style comment, such as `/* comment */`
+    CStyleComment {
+        /// The comment, ignoring initial `--`
+        comment: ShortString,
+    },
 }
 
 impl TokenType {
     /// Returns whether a token can be practically ignored in most cases
     /// Comments and whitespace will return `true`, everything else will return `false`
     pub fn is_trivia(&self) -> bool {
-        matches!(
+        #[cfg(not(feature = "cfxlua"))]
+        return matches!(
             self,
             TokenType::Shebang { .. }
                 | TokenType::SingleLineComment { .. }
                 | TokenType::MultiLineComment { .. }
                 | TokenType::Whitespace { .. }
-        )
+        );
+
+        #[cfg(feature = "cfxlua")]
+        return matches!(
+            self,
+            TokenType::Shebang { .. }
+                | TokenType::SingleLineComment { .. }
+                | TokenType::MultiLineComment { .. }
+                | TokenType::Whitespace { .. }
+                | TokenType::CStyleComment { .. }
+        );
     }
 
     /// Returns the kind of the token type.
@@ -347,6 +371,9 @@ impl TokenType {
 
             #[cfg(feature = "luau")]
             TokenType::InterpolatedString { .. } => TokenKind::InterpolatedString,
+
+            #[cfg(feature = "cfxlua")]
+            TokenType::CStyleComment { .. } => TokenKind::CStyleComment,
         }
     }
 
@@ -391,6 +418,10 @@ pub enum TokenKind {
     #[cfg(feature = "luau")]
     /// Some form of interpolated string
     InterpolatedString,
+
+    #[cfg(feature = "cfxlua")]
+    /// A C-style comment, such as `/* comment */`
+    CStyleComment,
 }
 
 /// A token such consisting of its [`Position`] and a [`TokenType`]
@@ -485,6 +516,9 @@ impl fmt::Display for Token {
                     write!(formatter, "`{literal}`")
                 }
             },
+
+            #[cfg(feature = "cfxlua")]
+            CStyleComment { comment } => write!(formatter, "/*{comment}*/"),
         }
     }
 }
@@ -518,6 +552,9 @@ impl Visit for Token {
 
             #[cfg(feature = "luau")]
             TokenKind::InterpolatedString => visitor.visit_interpolated_string_segment(self),
+
+            #[cfg(feature = "cfxlua")]
+            TokenKind::CStyleComment => visitor.visit_c_style_comment(self),
         }
     }
 }
@@ -539,6 +576,9 @@ impl VisitMut for Token {
 
             #[cfg(feature = "luau")]
             TokenKind::InterpolatedString => visitor.visit_interpolated_string_segment(token),
+
+            #[cfg(feature = "cfxlua")]
+            TokenKind::CStyleComment => visitor.visit_c_style_comment(token),
         }
     }
 }
@@ -857,6 +897,9 @@ pub enum StringLiteralQuoteType {
     Double,
     /// Strings formatted 'with single quotes'
     Single,
+    /// Strings in form of Jenkins Hash `with backticks`
+    #[cfg(feature = "cfxlua")]
+    Backtick,
 }
 
 impl fmt::Display for StringLiteralQuoteType {
@@ -868,6 +911,8 @@ impl fmt::Display for StringLiteralQuoteType {
             StringLiteralQuoteType::Brackets => Err(fmt::Error),
             StringLiteralQuoteType::Double => "\"".fmt(formatter),
             StringLiteralQuoteType::Single => "'".fmt(formatter),
+            #[cfg(feature = "cfxlua")]
+            StringLiteralQuoteType::Backtick => "`".fmt(formatter),
         }
     }
 }