+++ full-moon/tests/cfxlua_cases/fail/tokenizer/c_style_comments/source.lua
@@ -1,0 +1,2 @@
+print("Hello, World!") /* Comment
+local a = 2